@@ -1,19 +1,19 @@
 apply plugin: 'com.android.application'
 
+apply plugin: 'kotlin-android'
+
+apply plugin: 'kotlin-android-extensions'
+
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
+    buildToolsVersion "29.0.2"
     defaultConfig {
-<<<<<<< HEAD
-        applicationId "com.android.myapplication"
-        minSdkVersion 26
-=======
-        applicationId "com.aula"
-        minSdkVersion 23
->>>>>>> 877bbba44308fc4e9feac97bd21481a1ae0c7db5
-        targetSdkVersion 28
+        applicationId "com.example.myapplication"
+        minSdkVersion 15
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
@@ -25,17 +25,14 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-<<<<<<< HEAD
-    implementation 'com.android.support:cardview-v7:28.0.0'
-=======
->>>>>>> 877bbba44308fc4e9feac97bd21481a1ae0c7db5
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation"org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.core:core-ktx:1.1.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-<<<<<<< HEAD
-=======
-    implementation 'com.android.support:design:28.0.0'
->>>>>>> 877bbba44308fc4e9feac97bd21481a1ae0c7db5
+    androidTestImplementation 'androidx.test:runner:1.2.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'com.android.support:cardview-v7:29.0.0'
+
 }
