@@ -1,7 +1,27 @@
 package com.example.greenvoice;
 
-import android.support.v7.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+import androidx.appcompat.app.AppCompatActivity;
 
 public class AddUserActivity extends AppCompatActivity {
 
+    @Override
+    protected void onCreate(Bundle a) {
+        super.onCreate(a);
+        setContentView(R.layout.add_user_view);
+
+
+        Button bt = findViewById(R.id.botaoCadastrarUsuario);
+        bt.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(AddUserActivity.this, ActivityOpcao.class);
+                startActivity(intent);
+            }
+        });
+    }
 }
