@@ -2,9 +2,9 @@ package com.example.greenvoice;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
+import androidx.appcompat.app.AppCompatActivity;
 
 public class ActivityAnonimo extends AppCompatActivity {
 
@@ -19,7 +19,7 @@ public class ActivityAnonimo extends AppCompatActivity {
         bt.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(ActivityAnonimo.this, ListaDeOcorrenciasActivity.class);
+                Intent intent = new Intent(ActivityAnonimo.this, ActivityOpcao.class);
                 startActivity(intent);
             }
         });
