@@ -1,6 +1,38 @@
 package com.example.greenvoice;
 
-import android.support.v7.app.AppCompatActivity;
 
-public class AddOcorrenciaActivity extends AppCompatActivity {
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.*;
+import android.widget.AdapterView.OnItemSelectedListener;
+import androidx.appcompat.app.AppCompatActivity;
+
+public class AddOcorrenciaActivity extends AppCompatActivity implements OnItemSelectedListener {
+
+    private Spinner sp;
+
+    @Override
+    protected void onCreate(Bundle a) {
+        super.onCreate(a);
+        setContentView(R.layout.add_ocorrencia);
+
+        sp = (Spinner)findViewById(R.id.spinner_categorias);
+        ArrayAdapter<CharSequence> adapter = ArrayAdapter.createFromResource(this, R.array.categorias, android.R.layout.simple_spinner_item);
+        adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
+        sp.setAdapter(adapter);
+
+        sp.setOnItemSelectedListener(this);
+    }
+
+    @Override
+    public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
+        String text = parent.getItemAtPosition(position).toString();
+        Toast.makeText(parent.getContext(), text, Toast.LENGTH_SHORT).show();
+    }
+
+    @Override
+    public void onNothingSelected(AdapterView<?> adapterView) {
+
+    }
 }
