@@ -2,10 +2,10 @@ package com.example.greenvoice;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
+import androidx.appcompat.app.AppCompatActivity;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -20,7 +20,7 @@ public class MainActivity extends AppCompatActivity {
         bt.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                 Intent intent = new Intent(MainActivity.this, ListaDeOcorrenciasActivity.class);
+                 Intent intent = new Intent(MainActivity.this, ActivityOpcao.class);
                  startActivity(intent);
             }
         });
@@ -38,7 +38,7 @@ public class MainActivity extends AppCompatActivity {
         btAnonimo.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(MainActivity.this, ListaDeOcorrenciasActivity.class);
+                Intent intent = new Intent(MainActivity.this, ActivityOpcao.class);
                 startActivity(intent);
             }
         });
